test(refactor): rename tools.js -> utils.js
